move BundleReport to base
