+++ graphql_api/types/bundle_analysis/__init__.py
@@ -5,6 +5,7 @@ from .base import (
     bundle_data_bindable,
     bundle_load_time_bindable,
     bundle_module_bindable,
+    bundle_report_bindable,
     bundle_size_bindable,
 )
 from .comparison import (
@@ -15,7 +16,6 @@ from .comparison import (
 from .report import (
     bundle_analysis_report_bindable,
     bundle_analysis_report_result_bindable,
-    bundle_report_bindable,
 )
 
 bundle_analysis = ariadne_load_local_graphql(__file__, "base.graphql")
