+++ graphql_api/types/bundle_analysis/report.py
@@ -7,7 +7,6 @@ from services.bundle_analysis import BundleAnalysisReport, BundleReport
 
 bundle_analysis_report_result_bindable = UnionType("BundleAnalysisReportResult")
 bundle_analysis_report_bindable = ObjectType("BundleAnalysisReport")
-bundle_report_bindable = ObjectType("BundleReport")
 
 
 @bundle_analysis_report_result_bindable.type_resolver
@@ -35,18 +34,3 @@ def resolve_bundles(
     bundles_analysis_report: BundleAnalysisReport, info
 ) -> List[BundleReport]:
     return bundles_analysis_report.bundles
-
-
-@bundle_report_bindable.field("name")
-def resolve_name(bundle_report: BundleReport, info) -> str:
-    return bundle_report.bundle_name
-
-
-@bundle_report_bindable.field("sizeTotal")
-def resolve_size_total(bundle_report: BundleReport, info) -> int:
-    return bundle_report.size_total
-
-
-@bundle_report_bindable.field("loadTimeTotal")
-def resolve_load_time_total(bundle_report: BundleReport, info) -> float:
-    return bundle_report.load_time_total
